+++ Integrations/ProofpointTAP_v2/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
  - fixed fetch-incidents when last_fetch was greater than 1 hour
+ - fixed fetch-incidents when operation timed out
 
 ## [19.8.0] - 2019-08-06
   - Modified the fetch range for the first fetch to 1 hour (the Proofpoint TAP API maximum).